Ticket 1053: refactored by moving DbNames to a separate file within core
